backend/requirements.txt
composio-core==0.7.4
langchain-community==0.3.19
langchain-anthropic==0.3.8
langchain-weaviate==0.0.4
langchain-huggingface==0.1.2
langchain-core==0.3.49
langgraph==0.3.2
langsmith==0.3.13
python-dotenv==1.0.1
pydantic==2.10.6
pydantic[email]
pydantic[email,timezone]
polars==1.22.0
composio-langchain==0.7.4
langchain-groq==0.2.5
tavily-python==0.5.1
langgraph-checkpoint-sqlite==2.0.6
weaviate-client==4.11.3
python-magic
pypdf==5.4.0
pytesseract==0.3.13
pillow==11.1.0
matplotlib==3.10.1
networkx==3.4.2
asyncio==3.4.3
typing-extensions==4.13.0
sentence-transformers==4.0.1
pdf2image==1.17.0
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
streamlit==1.44.0
requests==2.32.3